Fix up pin data for trips routes
Prior to this, pins were not being added to trips properly because the POST/create route attempted to run the pin queries synchronously instead of async. Thus, the pins array on any given trip was never populated before the trip data was saved. This change reverses the order of the "join", grabbing pin data on demand when a trip is retrieved to avoid storing stale pin info in the trip itself.